confirmation-page: fix component splitting
splits both error and success as different components

also fixes the layout on payment page for boleto buttons
