+++ webapp/src/theme-pagarme/confirmation-page/properties.css
@@ -9,17 +9,7 @@
   --page-title-margin: 6px 0;
 
   --confirmation-font-family: var(--primary-font-family);
-  --confirmation-success: var(--color-success);
-  --confirmation-error: var(--color-error);
-  --confirmation-info: var(--font-size-large);
   --background-fill-color: var(--color-grey-83);
-
-  --field-color: var(--color-grey-39);
-  --field-margin-bottom: var(--spacing-tiny);
-  --field-font-size: var(--font-size-small);
-
-  --value-color: var(--color-white);
-  --value-font-size: var(--font-size-medium);
 
   --success-icon-padding-top: 10px;
   --error-icon-default-padding: 8px;
