+++ webapp/src/theme-pagarme/success-info/properties.css
@@ -1,0 +1,18 @@
+@import "../colors.css";
+@import "../typography.css";
+@import "../spacing.css";
+
+:root {
+  --page-title-font-size: var(--font-size-xxlarge);
+  --page-title-color: var(--color-white);
+
+  --confirmation-font-family: var(--primary-font-family);
+  --confirmation-success: var(--color-success);
+
+  --field-color: var(--color-grey-39);
+  --field-margin-bottom: var(--spacing-tiny);
+  --field-font-size: var(--font-size-small);
+
+  --value-color: var(--color-white);
+  --value-font-size: var(--font-size-medium);
+}
